feat: simplify package structure
