+++ packages/abaca-cli/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "abaca-cli",
+  "version": "0.11.0",
+  "repository": "github:opvious/abaca",
+  "description": "OpenAPI SDK generator with strong type guarantees and minimal boilerplate",
+  "keywords": [
+    "abaca",
+    "cli",
+    "fetch",
+    "openapi",
+    "sdk",
+    "typescript"
+  ],
+  "author": "Opvious Engineering <oss@opvious.io>",
+  "license": "Apache-2.0",
+  "type": "module",
+  "engines": {
+    "node": ">=20"
+  },
+  "files": [
+    "bin/",
+    "lib/",
+    "resources/templates"
+  ],
+  "bin": {
+    "abaca": "bin/abaca.js"
+  },
+  "exports": {},
+  "scripts": {
+    "build": "tsc -b src",
+    "clean": "rm -rf lib node_modules out $(find . -name '*.gen' -o -name '*.gen.*')",
+    "fix": "prettier --write",
+    "gen-preamble": "./scripts/generate-preamble-template",
+    "gen-test-pets": "./bin/abaca.js g test/resources/pets.openapi.yaml -o test/pets-sdk.gen.ts",
+    "gen-test-snippets": "./bin/abaca.js g test/resources/snippets.openapi.yaml -o test/snippets-sdk.gen.ts --generate-ids",
+    "gen-test-tables": "./bin/abaca.js g test/resources/tables.openapi.yaml -o test/tables-sdk.gen.ts --strict-additional-properties",
+    "gen-test": "pnpm run gen-test-pets && pnpm run gen-test-snippets && pnpm run gen-test-tables && tsc -b test",
+    "prepare": "pnpm run gen-preamble && pnpm run build && inlinable r",
+    "test": "pnpm run gen-test && vitest",
+    "watch": "tsc -b -w src test"
+  },
+  "dependencies": {
+    "@opentelemetry/api": "~1.8.0",
+    "@opvious/stl-bootstrap": "^0.25.4",
+    "@opvious/stl-errors": "^0.25.4",
+    "@opvious/stl-utils": "^0.25.4",
+    "abaca": "workspace:*",
+    "abaca-openapi": "workspace:*",
+    "commander": "^12.1.0",
+    "eta": "^3.4.0",
+    "node-fetch": "^3.3.2",
+    "openapi-typescript": "^6.7.6",
+    "ora": "^8.0.1",
+    "yaml": "^2.4.5"
+  },
+  "devDependencies": {
+    "@koa/router": "^12.0.1",
+    "@types/koa": "^2.15.0",
+    "@types/koa__router": "^12.0.4",
+    "inlinable": "^0.3.3",
+    "koa": "^2.15.3",
+    "koa-body": "^6.0.1",
+    "ts-essentials": "^9.4.2"
+  }
+}
