+++ packages/abaca-cli/resources/templates/default-factory.eta
