+++ packages/abaca-cli/resources/templates/smart-factory.eta
