+++ packages/abaca-cli/src/index.ts
@@ -1,0 +1,2 @@
+export {mainCommand} from './command/index.js';
+export {telemetry} from './common.js';
