+++ packages/abaca-cli/test/resources/pets.openapi.yaml
