+++ packages/abaca-cli/test/resources/snippets.openapi.yaml
