+++ packages/abaca-cli/test/resources/tables.openapi.yaml
