--- packages/abaca-codecs/README.md
@@ -1,7 +1,0 @@
-# Abaca SDK codecs
-
-Common codecs to use with client SDKs.
-
-+ Node-specific (requiring
-  [`node-fetch`](https://github.com/node-fetch/node-fetch)):
-  + [`application/json-seq`](src/node/json-seq.ts)
