--- packages/abaca-codecs/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "abaca-codecs",
-  "version": "0.11.0",
-  "repository": "github:opvious/abaca",
-  "description": "Abaca SDK codecs",
-  "keywords": [
-    "abaca"
-  ],
-  "author": "Opvious Engineering <oss@opvious.io>",
-  "license": "Apache-2.0",
-  "type": "module",
-  "files": [
-    "lib/"
-  ],
-  "exports": {
-    "./node/json-seq": "./lib/node/json-seq.js"
-  },
-  "scripts": {
-    "build": "tsc -p src",
-    "clean": "rm -rf lib node_modules out",
-    "fix": "prettier --write",
-    "prepare": "pnpm run build",
-    "watch": "tsc -b -w src"
-  },
-  "dependencies": {
-    "@opvious/stl-errors": "^0.25.4",
-    "abaca-runtime": "workspace:*",
-    "json-text-sequence": "^2.0.2",
-    "node-fetch": "^3.3.2"
-  }
-}
