+++ packages/abaca-koa/package.json
@@ -39,8 +39,8 @@
     "@opvious/stl-errors": "^0.25.4",
     "@opvious/stl-telemetry": "^0.25.4",
     "@opvious/stl-utils": "^0.25.4",
+    "abaca": "workspace:*",
     "abaca-openapi": "workspace:*",
-    "abaca-runtime": "workspace:*",
     "ajv": "^8.16.0",
     "busboy": "^1.6.0",
     "co-body": "^6.2.0",
@@ -59,7 +59,6 @@
     "@types/koa__router": "^12.0.4",
     "@types/qs": "^6.9.15",
     "abaca": "workspace:*",
-    "abaca-codecs": "workspace:*",
     "inlinable": "^0.3.3",
     "koa": "^2.15.3",
     "koa-body": "^6.0.1",
