+++ packages/abaca-openapi/package.json
@@ -30,7 +30,7 @@
     "@opvious/stl-telemetry": "^0.25.4",
     "@opvious/stl-utils": "^0.25.4",
     "@stoplight/json-ref-resolver": "^3.1.6",
-    "abaca-runtime": "workspace:*",
+    "abaca": "workspace:*",
     "ajv": "^8.16.0",
     "openapi-schema-validator": "^12.1.3",
     "openapi-types": "^12.1.3",
