--- packages/abaca-runtime/README.md
@@ -1,3 +1,0 @@
-# Abaca runtime utilities
-
-Internal utilities shared by the various Abaca packages.
