--- packages/abaca-runtime/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "abaca-runtime",
-  "version": "0.11.0",
-  "repository": "github:opvious/abaca",
-  "description": "Abaca runtime utilities",
-  "keywords": [
-    "abaca"
-  ],
-  "author": "Opvious Engineering <oss@opvious.io>",
-  "license": "Apache-2.0",
-  "type": "module",
-  "files": [
-    "lib/"
-  ],
-  "exports": "./lib/index.js",
-  "scripts": {
-    "build": "tsc -p src",
-    "clean": "rm -rf lib node_modules out",
-    "fix": "prettier --write",
-    "prepare": "pnpm run build",
-    "watch": "tsc -b -w src"
-  }
-}
