+++ packages/abaca/src/common.ts
@@ -1,24 +1,27 @@
-import {appTelemetry} from '@opvious/stl-bootstrap';
-import {ResourceLoader} from '@opvious/stl-utils/files';
-import __inlinable from 'inlinable';
-import os from 'os';
-import path from 'path';
+export type AsyncOrSync<V> = V | Promise<V>;
 
-export const packageInfo = __inlinable((ctx) =>
-  ctx.enclosing(import.meta.url).metadata()
-);
+export type Has<N extends number | string | symbol, V> = {
+  readonly [K in N]: V;
+};
 
-export const resourceLoader = ResourceLoader.enclosing(import.meta.url);
+export type Get<O, N extends keyof O | string, D = never> =
+  O extends Has<N, infer V> ? V & {} : D;
 
-export const COMMAND_NAME = packageInfo.name;
+export type Lookup<O, N extends keyof O | string, D = undefined> =
+  O extends Has<N, infer V>
+    ? V & {}
+    : O extends Partial<Has<N, infer V>>
+      ? (V & {}) | undefined
+      : D;
 
-export function logPath(): string {
-  return path.join(os.tmpdir(), packageInfo.name + '.log');
-}
+export type Exact<T, V> = T extends V
+  ? Exclude<keyof T, keyof V> extends never
+    ? T
+    : never
+  : never;
 
-export const telemetry = appTelemetry(packageInfo, {
-  loggerOptions: {
-    destination: logPath(),
-    base: {pid: process.pid},
-  },
-});
+export type Values<O> = O[keyof O];
+
+export type KeysOfValues<O> = Values<{
+  [K in keyof O]: O[K] extends never ? never : keyof O[K];
+}>;
